py_strategy/requirements.txt
cycler==0.11.0
flatbuffers==25.2.10
joblib==1.1.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
onnxruntime==1.10.0
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.6
pybind11==2.12.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.1
pyzmq==25.1.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.17.0
threadpoolctl==3.1.0
xgboost==1.5.2
